without_generative_video/requirements.txt
google-genai
google-api-python-client
google-auth
google-auth-oauthlib
moviepy
